WIP: server starts and endpoints look ok
