+++ python/sidecar/ai/ranker.py
