+++ python/sidecar/ai/rewrite.py
@@ -1,19 +1,76 @@
 import os
+import re
 
 import openai
-from sidecar import prompts, shared, typing
-from sidecar.settings import logger
-from sidecar.typing import (
+from sidecar.ai.prompts import (
+    create_prompt_for_rewrite,
+    create_prompt_for_text_completion,
+)
+from sidecar.ai.schemas import (
     RewriteChoice,
     RewriteRequest,
+    RewriteResponse,
     TextCompletionChoice,
     TextCompletionRequest,
+    TextCompletionResponse,
     TextSuggestionChoice,
 )
+from sidecar.config import logger
+from sidecar.settings.schemas import FlatSettingsSchema
+from sidecar.typing import ResponseStatus
 from tenacity import retry, stop_after_attempt, wait_fixed
 
 
-def rewrite(arg: RewriteRequest, user_settings: typing.FlatSettingsSchema = None):
+def trim_completion_prefix_from_choices(
+    prefix: str, choices: list[TextCompletionChoice]
+) -> list[TextCompletionChoice]:
+    """
+    Trim a text completion prefix from a list of text completion choices.
+
+    Parameters
+    ----------
+    prefix : str
+        The prefix to trim from the text completion choices.
+    choices : List[TextCompletionChoice]
+        The text completion choices to trim the prefix from.
+
+    Returns
+    -------
+    List[TextCompletionChoice]
+        The text completion choices with the prefix trimmed.
+
+    Examples
+    --------
+    >>> prefix = "This is a prefix and "
+    >>> choices = [
+    ...     TextCompletionChoice(index=0, text="This is a prefix and more text."),
+    ...     TextCompletionChoice(index=1, text="This is a prefix and other text."),
+    ...     TextCompletionChoice(index=2, text="some text."),
+    ... ]
+    >>> trimmed_choices = trim_completion_prefix_from_choices(prefix, choices)
+    >>> trimmed_choices[0].text
+    'more text.'
+    >>> trimmed_choices[1].text
+    'other text.'
+    >>> trimmed_choices[2].text
+    'some text.'
+    """
+    # split prefix into sentences
+    sentence_splitter_regex = r"(?<!\w\.\w.)(?<![A-Z][a-z]\.)(?<=\.|\?)\s"
+    prefix_sentences = re.split(sentence_splitter_regex, prefix)
+
+    # and remove any sentences from the prompt that were included in the responses
+    for choice in choices:
+        for sentence in prefix_sentences:
+            if choice.text.strip().startswith(sentence.strip()):
+                choice.text = choice.text[len(sentence) :].strip()
+
+            if "[MASK]" in choice.text:
+                choice.text = choice.text.replace("[MASK]", "")
+    return choices
+
+
+def rewrite(arg: RewriteRequest, user_settings: FlatSettingsSchema = None):
     text = arg.text
     manner = arg.manner
     n_choices = arg.n_choices
@@ -33,23 +90,23 @@ def rewrite(arg: RewriteRequest, user_settings: typing.FlatSettingsSchema = None
 
     logger.info(f"Calling rewrite with the following parameters: {arg.dict()}")
 
-    prompt = prompts.create_prompt_for_rewrite(text, manner)
+    prompt = create_prompt_for_rewrite(text, manner)
     chat = Rewriter(prompt, n_choices, temperature, max_tokens, model)
 
     try:
         choices = chat.get_response(response_type=RewriteChoice)
     except Exception as e:
         logger.error(e)
-        response = typing.RewriteResponse(
-            status=typing.ResponseStatus.ERROR,
+        response = RewriteResponse(
+            status=ResponseStatus.ERROR,
             message=str(e),
             choices=[],
         )
         return response
 
     logger.info(f"Returning {len(choices)} rewrite choices to client: {choices}")
-    response = typing.RewriteResponse(
-        status=typing.ResponseStatus.OK,
+    response = RewriteResponse(
+        status=ResponseStatus.OK,
         message="",
         choices=[r.dict() for r in choices],
     )
@@ -57,7 +114,7 @@ def rewrite(arg: RewriteRequest, user_settings: typing.FlatSettingsSchema = None
 
 
 def complete_text(
-    request: TextCompletionRequest, user_settings: typing.FlatSettingsSchema = None
+    request: TextCompletionRequest, user_settings: FlatSettingsSchema = None
 ):
     if user_settings is not None:
         openai.api_key = user_settings.openai_api_key
@@ -69,7 +126,7 @@ def complete_text(
     logger.info(
         f"Calling text completion with the following parameters: {request.dict()}"
     )
-    prompt = prompts.create_prompt_for_text_completion(request)
+    prompt = create_prompt_for_text_completion(request)
     chat = Rewriter(
         prompt,
         n_choices=request.n_choices,
@@ -82,21 +139,19 @@ def complete_text(
         choices = chat.get_response(response_type=TextCompletionChoice)
     except Exception as e:
         logger.error(e)
-        response = typing.TextCompletionResponse(
-            status=typing.ResponseStatus.ERROR,
+        response = TextCompletionResponse(
+            status=ResponseStatus.ERROR,
             message=str(e),
             choices=[],
         )
         return response
 
     logger.info(f"Trimming completion prefix text from completion choices: {choices}")
-    choices = shared.trim_completion_prefix_from_choices(
-        prefix=request.text, choices=choices
-    )
+    choices = trim_completion_prefix_from_choices(prefix=request.text, choices=choices)
     logger.info(f"Returning {len(choices)} completion choices to client: {choices}")
 
-    response = typing.TextCompletionResponse(
-        status=typing.ResponseStatus.OK,
+    response = TextCompletionResponse(
+        status=ResponseStatus.OK,
         message="",
         choices=[r.dict() for r in choices],
     )
@@ -171,6 +226,9 @@ class Rewriter:
 
     def prepare_messages_for_chat(self, text: str) -> list:
         messages = [
+            {"role": "user", "content": text},
+        ]
+        return messages
             {"role": "user", "content": text},
         ]
         return messages
