+++ python/sidecar/references/storage.py
@@ -1,27 +1,37 @@
 import json
 
-from sidecar import settings, typing
-from sidecar.settings import logger
-from sidecar.typing import DeleteRequest, ReferenceUpdate
+from sidecar import config
+from sidecar.config import logger
+from sidecar.references.schemas import (
+    Author,
+    Chunk,
+    DeleteRequest,
+    DeleteStatusResponse,
+    Reference,
+    ReferencePatch,
+    ReferenceUpdate,
+    UpdateStatusResponse,
+)
+from sidecar.typing import ResponseStatus
 
 logger = logger.getChild(__name__)
 
 
 def get_reference(reference_id: str):
-    storage = JsonStorage(settings.REFERENCES_JSON_PATH)
+    storage = JsonStorage(config.REFERENCES_JSON_PATH)
     storage.load()
     return storage.get_reference(reference_id)
 
 
 def update_reference(reference_id: str, reference_update: ReferenceUpdate):
-    storage = JsonStorage(settings.REFERENCES_JSON_PATH)
+    storage = JsonStorage(config.REFERENCES_JSON_PATH)
     storage.load()
     response = storage.update(reference_id, reference_update=reference_update)
     return response
 
 
 def delete_references(delete_request: DeleteRequest):
-    storage = JsonStorage(settings.REFERENCES_JSON_PATH)
+    storage = JsonStorage(config.REFERENCES_JSON_PATH)
     storage.load()
     response = storage.delete(ids=delete_request.reference_ids, all_=delete_request.all)
     return response
@@ -42,10 +52,10 @@ class JsonStorage:
         for item in data:
             for k, v in item.items():
                 if k == "authors":
-                    authors = [typing.Author(**a) for a in v]
+                    authors = [Author(**a) for a in v]
                 elif k == "chunks":
-                    chunks = [typing.Chunk(**c) for c in v]
-            ref = typing.Reference(**item)
+                    chunks = [Chunk(**c) for c in v]
+            ref = Reference(**item)
             ref.authors = authors
             ref.chunks = chunks
             self.references.append(ref)
@@ -59,7 +69,7 @@ class JsonStorage:
         with open(self.filepath, "w") as f:
             json.dump(contents, f, indent=2, default=str)
 
-    def get_reference(self, reference_id: str) -> typing.Reference | None:
+    def get_reference(self, reference_id: str) -> Reference | None:
         """
         Get a Reference from storage by id.
         """
@@ -98,20 +108,18 @@ class JsonStorage:
             except KeyError:
                 msg = f"Unable to delete {ref_id}: not found in storage"
                 logger.warning(msg)
-                response = typing.DeleteStatusResponse(
-                    status=typing.ResponseStatus.ERROR, message=msg
+                response = DeleteStatusResponse(
+                    status=ResponseStatus.ERROR, message=msg
                 )
                 return response
 
         self.references = list(refs.values())
         self.save()
 
-        response = typing.DeleteStatusResponse(
-            status=typing.ResponseStatus.OK, message=""
-        )
+        response = DeleteStatusResponse(status=ResponseStatus.OK, message="")
         return response
 
-    def update(self, reference_id: str, patch: typing.ReferencePatch):
+    def update(self, reference_id: str, patch: ReferencePatch):
         """
         Update a Reference in storage with the target reference.
         This is used when the client has updated the reference in the UI.
@@ -130,9 +138,7 @@ class JsonStorage:
         except KeyError:
             msg = f"Unable to update {reference_id}: not found in storage"
             logger.error(msg)
-            response = typing.UpdateStatusResponse(
-                status=typing.ResponseStatus.ERROR, message=msg
-            )
+            response = UpdateStatusResponse(status=ResponseStatus.ERROR, message=msg)
             return response
 
         logger.info(f"Updating {reference_id} with new values: {patch.data}")
@@ -141,9 +147,7 @@ class JsonStorage:
         self.references = list(refs.values())
         self.save()
 
-        response = typing.UpdateStatusResponse(
-            status=typing.ResponseStatus.OK, message=""
-        )
+        response = UpdateStatusResponse(status=ResponseStatus.OK, message="")
         return response
 
     def create_corpus(self):
